// backend/src/modules/auth/application/enums/auth-error-code.enum.ts
export enum AuthErrorCode {
  ACCOUNT_ALREADY_ACTIVATED = "ACCOUNT_ALREADY_ACTIVATED",
  ACTIVATION_TOKEN_EXPIRED = "ACTIVATION_TOKEN_EXPIRED",
  INVALID_ACTIVATION_TOKEN = "INVALID_ACTIVATION_TOKEN",
  INVALID_CREDENTIALS = "INVALID_CREDENTIALS",
  INVALID_OTP = "INVALID_OTP",
  INVALID_REFRESH_TOKEN = "INVALID_REFRESH_TOKEN",
  OTP_EXPIRED = "OTP_EXPIRED",
  REFRESH_TOKEN_EXPIRED = "REFRESH_TOKEN_EXPIRED",
  USER_NOT_ACTIVE = "USER_NOT_ACTIVE",
  USER_NOT_FOUND = "USER_NOT_FOUND",
}